Created component for map and cleaned App.js file
